exception use

+++ src/main/java/life/lyp/community/community/controller/PublishController.java
@@ -2,6 +2,7 @@ package life.lyp.community.community.controller;
 
 import life.lyp.community.community.dto.QuestionDTO;
 import life.lyp.community.community.model.Question;
+import life.lyp.community.community.model.User;
 import life.lyp.community.community.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -10,7 +11,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import life.lyp.community.community.model.User;
+
 import javax.servlet.http.HttpServletRequest;
 
 /**
@@ -28,7 +29,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -47,7 +48,7 @@ public class PublishController {
     @RequestParam("title") String title,
     @RequestParam("description") String description,
     @RequestParam("tag") String tag,
-    @RequestParam(value = "id",required = false) Integer id,
+    @RequestParam(value = "id",required = false) Long id,
     HttpServletRequest request,
     Model model
     ){
